chore(persistence) Abstract the persistence machine to support the visualiser
