+++ editor/src/components/editor/persistence/persistence.ts
@@ -1,0 +1,292 @@
+import {
+  actions,
+  assign,
+  createMachine,
+  DoneInvokeEvent,
+  interpret,
+  Interpreter,
+  send,
+} from 'xstate'
+import type { Model } from 'xstate/lib/model.types'
+import { ProjectFile } from '../../../core/shared/project-file-types'
+import { projectURLForProject } from '../../../core/shared/utils'
+import { defaultProject } from '../../../sample-projects/sample-project-utils'
+import { notice } from '../../common/notice'
+import { EditorAction, EditorDispatch } from '../action-types'
+import {
+  setForkedFromProjectID,
+  setProjectID,
+  setProjectName,
+  showToast,
+  updateFile,
+} from '../actions/action-creators'
+import { createNewProjectName, PersistentModel } from '../store/editor-state'
+import {
+  PersistenceContext,
+  PersistenceEvent,
+  SaveEvent,
+  createPersistenceMachine,
+  LoggedIn,
+  Forking,
+  CreatingProjectId,
+  Ready,
+  loadEvent,
+  newEvent,
+  saveEvent,
+  forkEvent,
+  userLogInEvent,
+  userLogOutEvent,
+} from './generic/persistence-machine'
+import {
+  PersistenceBackendAPI,
+  ProjectLoadResult,
+  ProjectModel,
+  ProjectWithFileChanges,
+} from './generic/persistence-types'
+
+export function pushProjectURLToBrowserHistory(projectId: string, projectName: string): void {
+  // Make sure we don't replace the query params
+  const queryParams = window.top.location.search
+  const projectURL = projectURLForProject(projectId, projectName)
+  const title = `Utopia ${projectName}`
+  window.top.history.pushState({}, title, `${projectURL}${queryParams}`)
+}
+
+export class PersistenceMachine {
+  private interpreter: Interpreter<
+    PersistenceContext<PersistentModel>,
+    any,
+    PersistenceEvent<PersistentModel, ProjectFile>
+  >
+  private lastSavedTS: number = 0
+  private throttledSaveTimeoutId: NodeJS.Timer | null = null
+  private waitingThrottledSaveEvent: SaveEvent<PersistentModel> | null = null
+  private queuedActions: Array<EditorAction> = [] // Queue up actions during events and transitions, then dispatch when ready
+
+  constructor(
+    backendAPI: PersistenceBackendAPI<PersistentModel, ProjectFile>,
+    dispatch: EditorDispatch,
+    onProjectNotFound: () => void,
+    onCreatedOrLoadedProject: (
+      projectId: string,
+      projectName: string,
+      project: PersistentModel,
+    ) => void,
+    private saveThrottle: number = 30000,
+  ) {
+    this.interpreter = interpret(
+      createPersistenceMachine<PersistentModel, ProjectFile>().withConfig({
+        services: {
+          getNewProjectId: backendAPI.getNewProjectId,
+          checkProjectOwned: (_, event) => {
+            if (event.type === 'BACKEND_CHECK_OWNERSHIP') {
+              return backendAPI.checkProjectOwned(event.projectId)
+            } else {
+              throw new Error(
+                `Incorrect event type triggered check ownership, ${JSON.stringify(event)}`,
+              )
+            }
+          },
+          downloadAssets: (_, event) => {
+            if (event.type === 'BACKEND_DOWNLOAD_ASSETS') {
+              return backendAPI.downloadAssets(event.projectId, event.projectModel)
+            } else {
+              throw new Error(
+                `Incorrect event type triggered asset download, ${JSON.stringify(event)}`,
+              )
+            }
+          },
+          saveProjectToServer: (context, event) => {
+            if (
+              event.type === 'BACKEND_SERVER_SAVE' &&
+              event.projectModel != null &&
+              context.projectId != null
+            ) {
+              return backendAPI.saveProjectToServer(context.projectId, event.projectModel)
+            } else {
+              throw new Error(
+                `Unable to save project with ID ${context.projectId} after event ${JSON.stringify(
+                  event,
+                )}`,
+              )
+            }
+          },
+          saveProjectLocally: (context, event) => {
+            if (
+              event.type === 'BACKEND_LOCAL_SAVE' &&
+              event.projectModel != null &&
+              context.projectId != null
+            ) {
+              return backendAPI.saveProjectLocally(context.projectId, event.projectModel)
+            } else {
+              throw new Error(
+                `Unable to save project with ID ${context.projectId} after event ${JSON.stringify(
+                  event,
+                )}`,
+              )
+            }
+          },
+          loadProject: (_, event) => {
+            if (event.type === 'BACKEND_LOAD') {
+              return backendAPI.loadProject(event.projectId)
+            } else {
+              throw new Error(`Invalid event type triggered project load ${JSON.stringify(event)}`)
+            }
+          },
+        },
+      }),
+    )
+
+    this.interpreter.onTransition((state, event) => {
+      if (state.changed) {
+        switch (event.type) {
+          case 'NEW_PROJECT_CREATED':
+            if (state.matches({ user: LoggedIn })) {
+              this.queuedActions.push(showToast(notice('Project successfully uploaded!')))
+            } else {
+              this.queuedActions.push(
+                showToast(notice('Locally cached project. Sign in to share!')),
+              )
+            }
+
+            onCreatedOrLoadedProject(
+              event.projectId,
+              event.projectModel.name,
+              event.projectModel.content,
+            )
+            break
+          case 'LOAD_COMPLETE':
+            onCreatedOrLoadedProject(
+              event.projectId,
+              event.projectModel.name,
+              event.projectModel.content,
+            )
+            break
+          case 'PROJECT_ID_CREATED':
+            this.queuedActions.push(setProjectID(event.projectId))
+            break
+          case 'LOAD_FAILED':
+            onProjectNotFound()
+            break
+          case 'DOWNLOAD_ASSETS_COMPLETE': {
+            if (state.matches({ core: { [Forking]: CreatingProjectId } })) {
+              this.queuedActions.push(setForkedFromProjectID(state.context.projectId!))
+              this.queuedActions.push(setProjectName(`${state.context.project!.name} (forked)`))
+              this.queuedActions.push(showToast(notice('Project successfully forked!')))
+            }
+
+            const updateFileActions = event.downloadAssetsResult.filesWithFileNames.map(
+              ({ fileName, file }) => updateFile(fileName, file, true),
+            )
+            this.queuedActions.push(...updateFileActions)
+            break
+          }
+          case 'SAVE_COMPLETE':
+            const updateFileActions = event.saveResult.filesWithFileNames.map(
+              ({ fileName, file }) => updateFile(fileName, file, true),
+            )
+            this.queuedActions.push(...updateFileActions)
+            this.lastSavedTS = Date.now()
+            // TODO Show toasts after first server save of a previously local project
+            break
+        }
+
+        if (state.matches({ core: Ready })) {
+          if (this.queuedActions.length > 0) {
+            const actionsToDispatch = this.queuedActions
+            const projectIdOrNameChanged = actionsToDispatch.some(
+              (action) =>
+                action.action === 'SET_PROJECT_ID' || action.action === 'SET_PROJECT_NAME',
+            )
+            if (projectIdOrNameChanged) {
+              pushProjectURLToBrowserHistory(state.context.projectId!, state.context.project!.name)
+            }
+            this.queuedActions = []
+            dispatch(actionsToDispatch)
+          }
+        }
+      }
+    })
+
+    this.interpreter.start()
+
+    window.addEventListener('beforeunload', (e) => {
+      if (!this.isSafeToClose()) {
+        this.sendThrottledSave()
+        e.preventDefault()
+        e.returnValue = ''
+      }
+    })
+  }
+
+  private isSafeToClose(): boolean {
+    return this.waitingThrottledSaveEvent == null
+  }
+
+  private clearThrottledSave(): void {
+    if (this.throttledSaveTimeoutId != null) {
+      clearTimeout(this.throttledSaveTimeoutId)
+      this.throttledSaveTimeoutId = null
+    }
+
+    this.waitingThrottledSaveEvent = null
+  }
+
+  private getRemainingSaveDelay(): number {
+    return Math.max(0, this.lastSavedTS + this.saveThrottle - Date.now())
+  }
+
+  private shouldThrottle(forceOrThrottle: 'force' | 'throttle'): boolean {
+    return forceOrThrottle === 'throttle' && this.getRemainingSaveDelay() > 0
+  }
+
+  // Exposed for testing
+  sendThrottledSave(): void {
+    if (this.waitingThrottledSaveEvent != null) {
+      this.interpreter.send(this.waitingThrottledSaveEvent)
+    }
+    this.clearThrottledSave()
+  }
+
+  // API
+
+  save(
+    projectName: string,
+    project: PersistentModel,
+    forceOrThrottle: 'force' | 'throttle' = 'throttle',
+  ): void {
+    const eventToFire = saveEvent({ name: projectName, content: project })
+
+    if (this.shouldThrottle(forceOrThrottle)) {
+      this.waitingThrottledSaveEvent = eventToFire
+      this.throttledSaveTimeoutId = setTimeout(this.sendThrottledSave, this.getRemainingSaveDelay())
+    } else {
+      this.interpreter.send(eventToFire)
+    }
+  }
+
+  load(projectId: string): void {
+    this.interpreter.send(loadEvent(projectId))
+  }
+
+  createNew(projectName: string, project: PersistentModel): void {
+    this.interpreter.send(newEvent({ name: projectName, content: project }))
+  }
+
+  fork(): void {
+    this.interpreter.send(forkEvent())
+  }
+
+  login(): void {
+    this.interpreter.send(userLogInEvent())
+  }
+
+  logout(): void {
+    this.interpreter.send(userLogOutEvent())
+  }
+
+  stop(): void {
+    this.interpreter.stop()
+    this.clearThrottledSave()
+  }
+}
